Remove V1 support and address the comments
