+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/sources/socket.scala
@@ -15,11 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.execution.streaming
+package org.apache.spark.sql.execution.streaming.sources
 
+import java.io.{BufferedReader, InputStreamReader, IOException}
+import java.net.Socket
 import java.sql.Timestamp
 import java.text.SimpleDateFormat
-import java.util.{List => JList, Locale, Optional}
+import java.util.{Calendar, List => JList, Locale, Optional}
+import javax.annotation.concurrent.GuardedBy
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable.ListBuffer
@@ -27,14 +30,12 @@ import scala.util.{Failure, Success, Try}
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql._
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.sources.{DataSourceRegister, StreamSourceProvider}
+import org.apache.spark.sql.sources.DataSourceRegister
 import org.apache.spark.sql.sources.v2.{DataSourceV2, DataSourceV2Options}
 import org.apache.spark.sql.sources.v2.reader.{DataReader, ReadTask}
 import org.apache.spark.sql.sources.v2.streaming.MicroBatchReadSupport
 import org.apache.spark.sql.sources.v2.streaming.reader.{MicroBatchReader, Offset => V2Offset}
 import org.apache.spark.sql.types.{StringType, StructField, StructType, TimestampType}
-import org.apache.spark.unsafe.types.UTF8String
 
 object TextSocketSource {
   val SCHEMA_REGULAR = StructType(StructField("value", StringType) :: Nil)
@@ -43,68 +44,83 @@ object TextSocketSource {
   val DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.US)
 }
 
-/**
- * A source that reads text lines through a TCP socket, designed only for tutorials and debugging.
- * This source will *not* work in production applications due to multiple reasons, including no
- * support for fault recovery and keeping all of the text read in memory forever.
- */
-class TextSocketSource(
-    protected val host: String,
-    protected val port: Int,
-    includeTimestamp: Boolean,
-    sqlContext: SQLContext)
-  extends Source with TextSocketReader with Logging {
-
-  initialize()
-
-  /** Returns the schema of the data from this source */
-  override def schema: StructType =
-    if (includeTimestamp) TextSocketSource.SCHEMA_TIMESTAMP else TextSocketSource.SCHEMA_REGULAR
-
-  override def getOffset: Option[Offset] = getOffsetInternal.map(LongOffset(_))
-
-  /** Returns the data that is between the offsets (`start`, `end`]. */
-  override def getBatch(start: Option[Offset], end: Offset): DataFrame = {
-    val rawList = getBatchInternal(start.flatMap(LongOffset.convert).map(_.offset),
-      LongOffset.convert(end).map(_.offset))
-
-    val rdd = sqlContext.sparkContext
-      .parallelize(rawList)
-      .map { case (v, ts) => InternalRow(UTF8String.fromString(v), ts.getTime) }
-    sqlContext.internalCreateDataFrame(rdd, schema, isStreaming = true)
-  }
-
-  override def commit(end: Offset): Unit = {
-    val newOffset = LongOffset.convert(end).getOrElse(
-      sys.error(s"TextSocketStream.commit() received an offset ($end) that did not " +
-        s"originate with an instance of this class")
-    )
-
-    commitInternal(newOffset.offset)
-  }
-
-  override def toString: String = s"TextSocketSource[host: $host, port: $port]"
-}
-
 case class TextSocketOffset(offset: Long) extends V2Offset {
   override def json(): String = offset.toString
 }
 
+/**
+ * A MicroBatchReader that reads text lines through a TCP socket, designed only for tutorials and
+ * debugging. This MicroBatchReader will *not* work in production applications due to multiple
+ * reasons, including no support for fault recovery and keeping all of the text read in memory
+ * forever.
+ */
 class TextSocketMicroBatchReader(options: DataSourceV2Options)
-  extends MicroBatchReader with TextSocketReader with Logging {
+  extends MicroBatchReader with Logging {
 
   private var startOffset: TextSocketOffset = _
   private var endOffset: TextSocketOffset = _
 
-  protected val host: String = options.get("host").get()
-  protected val port: Int = options.get("port").get().toInt
+  private val host: String = options.get("host").get()
+  private val port: Int = options.get("port").get().toInt
+
+  @GuardedBy("this")
+  private var socket: Socket = null
+
+  @GuardedBy("this")
+  private var readThread: Thread = null
+
+  /**
+   * All batches from `lastCommittedOffset + 1` to `currentOffset`, inclusive.
+   * Stored in a ListBuffer to facilitate removing committed batches.
+   */
+  @GuardedBy("this")
+  private val batches = new ListBuffer[(String, Timestamp)]
+
+  @GuardedBy("this")
+  private var currentOffset: Long = -1L
+
+  @GuardedBy("this")
+  private var lastOffsetCommitted: Long = -1L
 
   initialize()
 
-  override def setOffsetRange(start: Optional[V2Offset], end: Optional[V2Offset]): Unit = {
+  private def initialize(): Unit = synchronized {
+    socket = new Socket(host, port)
+    val reader = new BufferedReader(new InputStreamReader(socket.getInputStream))
+    readThread = new Thread(s"TextSocketSource($host, $port)") {
+      setDaemon(true)
+
+      override def run(): Unit = {
+        try {
+          while (true) {
+            val line = reader.readLine()
+            if (line == null) {
+              // End of file reached
+              logWarning(s"Stream closed by $host:$port")
+              return
+            }
+            TextSocketMicroBatchReader.this.synchronized {
+              val newData = (line,
+                Timestamp.valueOf(
+                  TextSocketSource.DATE_FORMAT.format(Calendar.getInstance().getTime()))
+              )
+              currentOffset += 1
+              batches.append(newData)
+            }
+          }
+        } catch {
+          case e: IOException =>
+        }
+      }
+    }
+    readThread.start()
+  }
+
+  override def setOffsetRange(
+      start: Optional[V2Offset],
+      end: Optional[V2Offset]): Unit = synchronized {
     startOffset = start.orElse(TextSocketOffset(-1L)).asInstanceOf[TextSocketOffset]
-    endOffset = end.orElse(
-      TextSocketOffset(getOffsetInternal.getOrElse(-1L))).asInstanceOf[TextSocketOffset]
+    endOffset = end.orElse(TextSocketOffset(currentOffset)).asInstanceOf[TextSocketOffset]
   }
 
   override def getStartOffset(): V2Offset = {
@@ -125,7 +141,18 @@ class TextSocketMicroBatchReader(options: DataSourceV2Options)
   }
 
   override def createReadTasks(): JList[ReadTask[Row]] = {
-    val rawList = getBatchInternal(Option(startOffset.offset), Option(endOffset.offset))
+    assert(startOffset != null && endOffset != null,
+      "start offset and end offset should already be set before create read tasks.")
+
+    val startOrdinal = startOffset.offset.toInt + 1
+    val endOrdinal = endOffset.offset.toInt + 1
+
+    // Internal buffer only holds the batches after lastOffsetCommitted
+    val rawList = synchronized {
+      val sliceStart = startOrdinal - lastOffsetCommitted.toInt - 1
+      val sliceEnd = endOrdinal - lastOffsetCommitted.toInt - 1
+      batches.slice(sliceStart, sliceEnd)
+    }
 
     assert(SparkSession.getActiveSession.isDefined)
     val spark = SparkSession.getActiveSession.get
@@ -157,16 +184,39 @@ class TextSocketMicroBatchReader(options: DataSourceV2Options)
     }.toList.asJava
   }
 
-  override def commit(end: V2Offset): Unit = {
+  override def commit(end: V2Offset): Unit = synchronized {
     val newOffset = end.asInstanceOf[TextSocketOffset]
-    commitInternal(newOffset.offset)
+
+    val offsetDiff = (newOffset.offset - lastOffsetCommitted).toInt
+
+    if (offsetDiff < 0) {
+      sys.error(s"Offsets committed out of order: $lastOffsetCommitted followed by $end")
+    }
+
+    batches.trimStart(offsetDiff)
+    lastOffsetCommitted = newOffset.offset
+  }
+
+  /** Stop this source. */
+  def stop(): Unit = synchronized {
+    if (socket != null) {
+      try {
+        // Unfortunately, BufferedReader.readLine() cannot be interrupted, so the only way to
+        // stop the readThread is to close the socket.
+        socket.close()
+      } catch {
+        case e: IOException =>
+      }
+      socket = null
+    }
   }
 
   override def toString: String = s"TextSocketMicroBatchReader[host: $host, port: $port]"
 }
 
 class TextSocketSourceProvider extends DataSourceV2
-  with MicroBatchReadSupport with StreamSourceProvider with DataSourceRegister with Logging {
+  with MicroBatchReadSupport with DataSourceRegister with Logging {
+
   private def checkParameters(params: Map[String, String]): Unit = {
     logWarning("The socket source should not be used for production applications! " +
       "It does not support recovery.")
@@ -186,37 +236,6 @@ class TextSocketSourceProvider extends DataSourceV2
       case Failure(_) =>
         throw new AnalysisException("includeTimestamp must be set to either \"true\" or \"false\"")
     }
-  }
-
-  /** Returns the name and schema of the source that can be used to continually read data. */
-  override def sourceSchema(
-      sqlContext: SQLContext,
-      schema: Option[StructType],
-      providerName: String,
-      parameters: Map[String, String]): (String, StructType) = {
-    checkParameters(parameters)
-    if (schema.nonEmpty) {
-      throw new AnalysisException("The socket source does not support a user-specified schema.")
-    }
-
-    val sourceSchema = if (parameters.getOrElse("includeTimestamp", "false").toBoolean) {
-        TextSocketSource.SCHEMA_TIMESTAMP
-      } else {
-        TextSocketSource.SCHEMA_REGULAR
-      }
-    ("textSocket", sourceSchema)
-  }
-
-  override def createSource(
-      sqlContext: SQLContext,
-      metadataPath: String,
-      schema: Option[StructType],
-      providerName: String,
-      parameters: Map[String, String]): Source = {
-    val host = parameters("host")
-    val port = parameters("port").toInt
-    new TextSocketSource(
-      host, port, parameters.getOrElse("includeTimestamp", "false").toBoolean, sqlContext)
   }
 
   override def createMicroBatchReader(
